+++ src/main/java/com/github/poxiton/utils/CommandCompletion.java
@@ -1,4 +1,4 @@
-package com.github.poxiton.commands;
+package com.github.poxiton.utils;
 
 import java.util.Arrays;
 import java.util.List;
@@ -7,7 +7,7 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
 
-public class DummyCompletion implements TabCompleter {
+public class CommandCompletion implements TabCompleter {
 
   @Override
   public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
